removed v1 types. removed some duplicated types between types and interfaces files
