+++ frontend/actions/review.ts
@@ -1,10 +1,9 @@
 import qs from 'querystring'
 import { ResponseTypeKeys } from 'src/common/ReviewWithComment'
 import useSWR from 'swr'
-import { ModelInterface, ReleaseInterface } from 'types/types'
-import { ReviewRequestInterface } from 'types/v2/types'
+import { ReleaseInterface } from 'types/types'
 
-import { AccessRequestInterface } from '../types/interfaces'
+import { AccessRequestInterface, ModelInterface, ReviewRequestInterface } from '../types/interfaces'
 import { ErrorInfo, fetcher } from '../utils/fetcher'
 
 export function useGetReviewRequestsForUser() {
