+++ frontend/src/common/MultiFileInput.tsx
@@ -4,7 +4,7 @@ import { styled } from '@mui/material/styles'
 import { ChangeEvent, useCallback, useMemo } from 'react'
 import MultiFileInputFileDisplay from 'src/common/MultiFileInputFileDisplay'
 import { FileWithMetadata } from 'types/interfaces'
-import { FileInterface } from 'types/v2/types'
+import { FileInterface } from 'types/types'
 
 const Input = styled('input')({
   display: 'none',
