+++ frontend/src/common/MultiFileInputFileDisplay.tsx
@@ -2,7 +2,7 @@ import { Chip, Grid, TextField, Tooltip, Typography } from '@mui/material'
 import prettyBytes from 'pretty-bytes'
 import { ChangeEvent, useState } from 'react'
 import { FileWithMetadata } from 'types/interfaces'
-import { FileInterface } from 'types/v2/types'
+import { FileInterface } from 'types/types'
 
 interface MultiFileInputDisplayProps {
   file: File | FileInterface
