+++ frontend/src/model/accessRequests/AccessRequestDisplay.tsx
@@ -4,8 +4,7 @@ import { groupBy } from 'lodash-es'
 import { useEffect, useState } from 'react'
 import UserDisplay from 'src/common/UserDisplay'
 import Link from 'src/Link'
-import { AccessRequestInterface } from 'types/interfaces'
-import { ReviewRequestInterface, ReviewResponse } from 'types/v2/types'
+import { AccessRequestInterface, ReviewRequestInterface, ReviewResponse } from 'types/interfaces'
 import { formatDateString, sortByCreatedAtAscending } from 'utils/dateUtils'
 import { plural } from 'utils/stringUtils'
