+++ frontend/src/model/overview/FormEditPage.tsx
@@ -6,8 +6,7 @@ import UnsavedChangesContext from 'src/contexts/unsavedChangesContext'
 import { useGetModel } from '../../../actions/model'
 import { putModelCard, useGetModelCardRevisions } from '../../../actions/modelCard'
 import { useGetSchema } from '../../../actions/schema'
-import { SplitSchemaNoRender } from '../../../types/interfaces'
-import { ModelInterface } from '../../../types/v2/types'
+import { ModelInterface, SplitSchemaNoRender } from '../../../types/interfaces'
 import { getStepsData, getStepsFromSchema } from '../../../utils/formUtils'
 import Loading from '../../common/Loading'
 import JsonSchemaForm from '../../Form/JsonSchemaForm'
