+++ frontend/src/model/overview/TemplatePage.tsx
@@ -2,7 +2,7 @@ import { PostAdd } from '@mui/icons-material'
 import { Box, Button, Card, Divider, Stack, Typography } from '@mui/material'
 import Link from 'src/Link'
 
-import { ModelInterface } from '../../../types/v2/types'
+import { ModelInterface } from '../../../types/interfaces'
 
 type TemplatePageProps = {
   model: ModelInterface
