+++ frontend/src/model/releases/ReleaseDisplay.tsx
@@ -6,12 +6,11 @@ import { useRouter } from 'next/router'
 import prettyBytes from 'pretty-bytes'
 import { useEffect, useState } from 'react'
 import UserDisplay from 'src/common/UserDisplay'
-import { ReviewRequestInterface, ReviewResponse } from 'types/v2/types'
 import { formatDateString, sortByCreatedAtAscending } from 'utils/dateUtils'
 
 import { useGetReviewRequestsForModel } from '../../../actions/review'
+import { ModelInterface, ReviewRequestInterface, ReviewResponse } from '../../../types/interfaces'
 import { ReleaseInterface } from '../../../types/types'
-import { ModelInterface } from '../../../types/v2/types'
 import Loading from '../../common/Loading'
 import Markdown from '../../common/MarkdownDisplay'
 import Link from '../../Link'
