+++ frontend/src/model/releases/ReleaseForm.tsx
@@ -11,8 +11,8 @@ import ReadOnlyAnswer from 'src/Form/ReadOnlyAnswer'
 import Link from 'src/Link'
 import MessageAlert from 'src/MessageAlert'
 import ModelImageList from 'src/model/ModelImageList'
-import { FileWithMetadata, FlattenedModelImage } from 'types/interfaces'
-import { FileInterface, isFileInterface, ModelInterface } from 'types/v2/types'
+import { FileWithMetadata, FlattenedModelImage, ModelInterface } from 'types/interfaces'
+import { FileInterface, isFileInterface } from 'types/types'
 import { isValidSemver } from 'utils/stringUtils'
 
 type ReleaseFormData = {
