+++ frontend/src/model/reviews/ApprovalsDisplay.tsx
@@ -5,7 +5,7 @@ import { useGetModelRoles } from 'actions/model'
 import { useMemo } from 'react'
 import Loading from 'src/common/Loading'
 import MessageAlert from 'src/MessageAlert'
-import { ReviewResponseWithRole } from 'types/v2/types'
+import { ReviewResponseWithRole } from 'types/interfaces'
 import { plural } from 'utils/stringUtils'
 
 interface ApprovalsDisplayProps {
