+++ frontend/src/model/settings/ModelDetails.tsx
@@ -6,10 +6,9 @@ import { FormEvent, useEffect, useState } from 'react'
 import ModelDescriptionInput from 'src/model/ModelDescriptionInput'
 import ModelNameInput from 'src/model/ModelNameInput'
 import TeamSelect from 'src/TeamSelect'
-import { TeamInterface } from 'types/interfaces'
+import { ModelForm, ModelInterface, TeamInterface } from 'types/interfaces'
 
 import { patchModel } from '../../../actions/model'
-import { ModelForm, ModelInterface } from '../../../types/v2/types'
 import { getErrorMessage } from '../../../utils/fetcher'
 import useNotification from '../../hooks/useNotification'
 import MessageAlert from '../../MessageAlert'
