+++ frontend/src/reviews/ReviewItem.tsx
@@ -2,7 +2,7 @@ import { ListItem, ListItemButton, Stack, Typography } from '@mui/material'
 import { useRouter } from 'next/router'
 import ReviewDisplay from 'src/model/reviews/ReviewDisplay'
 import ReviewRoleDisplay from 'src/reviews/ReviewRoleDisplay'
-import { ReviewRequestInterface } from 'types/v2/types'
+import { ReviewRequestInterface } from 'types/interfaces'
 import { timeDifference } from 'utils/dateUtils'
 
 type ReviewItemProps = {
