+++ frontend/src/settings/authentication/AuthenticationTab.tsx
@@ -8,7 +8,7 @@ import ConfirmationDialogue from 'src/common/ConfirmationDialogue'
 import EmptyBlob from 'src/common/EmptyBlob'
 import Loading from 'src/common/Loading'
 import MessageAlert from 'src/MessageAlert'
-import { TokenInterface } from 'types/v2/types'
+import { TokenInterface } from 'types/types'
 import { formatDateString } from 'utils/dateUtils'
 import { getErrorMessage } from 'utils/fetcher'
