+++ frontend/src/wrapper/TopNavigation.tsx
@@ -27,8 +27,7 @@ import { CSSProperties, MouseEvent, useContext, useMemo, useState } from 'react'
 import ModelSearchField from 'src/wrapper/ModelSearchField'
 
 import bailoLogo from '../../public/logo-horizontal-light.png'
-import { EntityKind } from '../../types/types'
-import { User } from '../../types/v2/types'
+import { EntityKind, User } from '../../types/types'
 import { DRAWER_WIDTH } from '../../utils/constants'
 import ExpandableButton from '../common/ExpandableButton'
 import UserAvatar from '../common/UserAvatar'
