--- frontend/types/v2/enums.ts
@@ -1,4 +1,0 @@
-export enum ApprovalStates {
-  Accepted = 'approve',
-  RequestChanges = 'request_changes',
-}
