--- frontend/types/v2/types.ts
@@ -1,164 +1,0 @@
-// TODO Once beta has been completed these types need to be merged back into types/types.
-// Please note that some of these types have been duplicated, merge accordingly!
-
-export const ModelVisibility = {
-  Private: 'private',
-  Public: 'public',
-} as const
-
-export type ModelVisibilityKeys = (typeof ModelVisibility)[keyof typeof ModelVisibility]
-
-export interface ModelInterface {
-  id: string
-  name: string
-  teamId: string
-  description: string
-  settings: {
-    ungovernedAccess: boolean
-  }
-  card: ModelCardInterface
-  visibility: ModelVisibilityKeys
-  collaborators: CollaboratorEntry[]
-}
-
-export interface ModelCardInterface {
-  schemaId: string
-  version: number
-  createdBy: string
-
-  metadata: unknown
-}
-
-export interface ModelCardRevisionInterface {
-  modelId: string
-  schemaId: string
-  version: number
-  metadata: unknown
-  createdBy: string
-  createdAt: string
-  updatedAt: string
-}
-
-export interface CollaboratorEntry {
-  entity: string
-  roles: Array<'owner' | 'contributor' | 'consumer' | string>
-}
-
-export type ModelForm = {
-  name: string
-  teamId: string
-  description: string
-  visibility: ModelVisibilityKeys
-}
-
-export interface Role {
-  id: string
-  name: string
-  short?: string
-}
-
-export const SchemaKind = {
-  Model: 'model',
-  AccessRequest: 'accessRequest',
-} as const
-
-export type SchemaKindKeys = (typeof SchemaKind)[keyof typeof SchemaKind]
-
-export interface FileInterface {
-  _id: string
-  modelId: string
-
-  name: string
-  size: number
-  mime: string
-
-  bucket: string
-  path: string
-
-  complete: boolean
-
-  createdAt: Date
-  updatedAt: Date
-}
-
-export const isFileInterface = (file: File | FileInterface): file is FileInterface => {
-  return (file as FileInterface).bucket !== undefined
-}
-
-export interface PostSimpleUpload {
-  file: FileInterface
-}
-
-export interface User {
-  dn: string
-}
-
-export interface EntityObject {
-  kind: string
-  id: string
-}
-
-export const TokenScope = {
-  All: 'all',
-  Models: 'models',
-} as const
-
-export type TokenScopeKeys = (typeof TokenScope)[keyof typeof TokenScope]
-
-export const TokenActions = {
-  ImageRead: 'image:read',
-  FileRead: 'file:read',
-} as const
-
-export type TokenActionsKeys = (typeof TokenActions)[keyof typeof TokenActions]
-
-export interface TokenInterface {
-  user: string
-  description: string
-  scope: TokenScopeKeys
-  modelIds: Array<string>
-  actions: Array<TokenActionsKeys>
-  accessKey: string
-  secretKey?: string
-  deleted: boolean
-  createdAt: string
-  updatedAt: string
-  compareToken: (candidateToken: string) => Promise<boolean>
-}
-
-export const Decision = {
-  RequestChanges: 'request_changes',
-  Approve: 'approve',
-} as const
-export type DecisionKeys = (typeof Decision)[keyof typeof Decision]
-
-export interface ReviewResponse {
-  user: string
-  decision: DecisionKeys
-  comment?: string
-  createdAt: string
-  updatedAt: string
-}
-
-export interface ReviewResponseWithRole extends ReviewResponse {
-  role: string
-}
-
-type PartialReviewRequestInterface =
-  | {
-      accessRequestId: string
-      semver?: never
-    }
-  | {
-      accessRequestId?: never
-      semver: string
-    }
-
-export type ReviewRequestInterface = {
-  model: ModelInterface
-  role: string
-  kind: 'release' | 'access'
-  responses: ReviewResponse[]
-  createdAt: string
-  updatedAt: string
-} & PartialReviewRequestInterface
